mspace and plot_mspace was modified so the rotation matrix selected for 3D data in the former could be inherited downstream to the latter
Adjustments in mspace and proj_landscape were made to deal with changing plot frames.

proj_consensus was deleted altogether, and proj_axis was modified so axes could be represented using arrows.

proj_phylogeny was modified so it no longer depends on other functions to compute shapes that will be used as tips.

a print.mspace function was also included to print mspace objects more amicably.

plot_mspace was modified so scatterpoints and hulls/ellipses from different groups are intercalated when the morphospace is regenerated (also changed to accommodate all the changes in the rest of the functions).

Other minor corrections and adjustments across the file were also included, and mspace workflow functions' documentation was updated.